+++ derive/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- `regex` dependency due to its
+  - Small usage within this crate
+  - Many dependencies, which causes long compile times
 
 ## 1.1.0 - 2017-09-18
 ### Added